@@ -1,8 +1,8 @@
-// This file is needed because it is used by vscode and other tools that
-// call `jest` directly.  However, unless you are doing anything special
-// do not edit this file
+// force timezone to UTC to allow tests to work regardless of local timezone
+// generally used by snapshots, but can affect specific tests
+process.env.TZ = 'UTC'
 
-const standard = require('@grafana/toolkit/src/config/jest.plugin.config');
-
-// This process will use the same config that `yarn test` is using
-module.exports = standard.jestConfig();
+module.exports = {
+  // Jest configuration provided by Grafana scaffolding
+  ...require('./.config/jest.config'),
+}
