@@ -1,10 +1,11 @@
 import { TimeRange } from '@grafana/data'
-import { InlineField, InlineFieldRow, RadioButtonGroup } from '@grafana/ui'
+import { Field, InlineField, InlineFieldRow, RadioButtonGroup } from '@grafana/ui'
 import { DataSource } from 'datasource'
 import defaults from 'lodash/defaults'
 import React, { useState } from 'react'
 import { defaultQuery, Query } from '../types'
 import { PathField } from './fields/PathField'
+import { QueryParamsEditor } from './fields/QueryParamsEditor'
 
 interface Props {
   onChange: (query: Query) => void
@@ -17,12 +18,7 @@ interface Props {
   fieldsTab: React.ReactNode
 }
 
-export const TabbedQueryEditor = ({
-  query,
-  onChange,
-  onRunQuery,
-  fieldsTab,
-}: Props) => {
+export const TabbedQueryEditor = ({ query, onChange, onRunQuery, fieldsTab }: Props) => {
   const [tabIndex, setTabIndex] = useState(0)
 
   const q = defaults(query, defaultQuery)
@@ -40,9 +36,20 @@ export const TabbedQueryEditor = ({
     {
       title: 'Path',
       content: (
-        <InlineField label='Path' tooltip='Websocket URL Path to connect.'>
-          <PathField path={q.path} onChange={onChangePath} />
-        </InlineField>
+        <>
+          <Field label='Path' description='Websocket URL path to connect'>
+            <PathField path={q.path} onChange={onChangePath} />
+          </Field>
+          <Field label='Query parameters' description='Optional query string parameters for the websocket connection'>
+            <QueryParamsEditor
+              values={Array.isArray(q.queryParams) ? q.queryParams : Object.entries(q.queryParams || {})}
+              onChange={qp => {
+                onChange({ ...q, queryParams: qp })
+                onRunQuery()
+              }}
+            />
+          </Field>
+        </>
       ),
     },
   ]
