@@ -1,9 +1,7 @@
 import { useEffect, useState } from 'react'
 
 export const useDebounce = <T,>(value: T, delay: number, initialValue?: T) => {
-  const [debouncedValue, setDebouncedValue] = useState<T | undefined>(
-    initialValue,
-  )
+  const [debouncedValue, setDebouncedValue] = useState<T | undefined>(initialValue)
 
   useEffect(() => {
     const handler = setTimeout(() => {
