@@ -1,4 +1,5 @@
-import { DataQuery, DataSourceJsonData, FieldType } from '@grafana/data'
+import { DataSourceJsonData, FieldType } from '@grafana/data'
+import { DataQuery } from '@grafana/schema'
 
 export type QueryLanguage = 'jsonpath'
 
@@ -13,14 +14,14 @@ export type Pair<T, K> = [T, K]
 
 export interface Query extends DataQuery {
   path: string
-  withStreaming: boolean
   fields: QueryField[]
+  queryParams?: Record<string, string> | Array<Pair<string, string>>
 }
 
 export const defaultQuery: Partial<Query> = {
   path: '',
   fields: [{ jsonPath: '', language: 'jsonpath', name: '' }],
-  withStreaming: true,
+  queryParams: [],
 }
 
 export interface DataSourceOptions extends DataSourceJsonData {
